moved background processing items to own module w/ burnettk
